+++ src/myinf/db.nim
@@ -1,17 +1,10 @@
 import myinf/configs
 
-import std/[
-  options,
-  logging
-]
-
 import pkg/norm/sqlite
 export sqlite
 
 import std/locks
 export locks
-
-addHandler newConsoleLogger(fmtStr = "")
 
 var dbLock*: Lock
 initLock dbLock
@@ -23,7 +16,6 @@ let dbConn* {.guard: dbLock.} = open(
 )
 
 template inDb*(body: untyped) =
-  # {.gcsafe.}:
-  block:
+  {.gcsafe.}:
     withLock dbLock:
       body