+++ src/myinf/routes.nim
@@ -1,6 +1,7 @@
 import pkg/prologue
 
-import myinf/routes/helloWorld
+import myinf/routes/home
+import myinf/routes/api/auth
 
 type Route = tuple
   path: string
@@ -8,6 +9,10 @@ type Route = tuple
 
 const routesDefinition*: seq[Route] = @[
   ("", @[
-    pattern("/", r_helloWorld, HttpGet, "Hello World"),
+    pattern("/", r_home, HttpGet, "Homepage"),
+  ]),
+  ("api", @[
+    pattern("/signin", r_signIn, HttpPost, "Sign In API"),
+    pattern("/signup", r_signUp, HttpPost, "Sign Up API"),
   ]),
 ]